Extract tenant-scoped warranty coverage service
